feat(api): separater raw view logic and openning
