+++ Holovibes/sources/gui/windows/panels/export_panel.cc
@@ -163,7 +163,7 @@ QString ExportPanel::browse_record_output_file()
     // Convert QString to std::string
     std::string std_filepath = filepath.toStdString();
 
-    const std::string file_ext = api::browse_record_output_file(std_filepath);
+    const std::string file_ext = gui::browse_record_output_file(std_filepath);
     // Will pick the item combobox related to file_ext if it exists, else, nothing is done
     ui_->RecordExtComboBox->setCurrentText(file_ext.c_str());
 
@@ -174,7 +174,7 @@ QString ExportPanel::browse_record_output_file()
 
 void ExportPanel::set_output_file_name(std::string std_filepath)
 {
-    const std::string file_ext = api::browse_record_output_file(std_filepath);
+    const std::string file_ext = gui::browse_record_output_file(std_filepath);
     // Will pick the item combobox related to file_ext if it exists, else, nothing is done
     ui_->RecordExtComboBox->setCurrentText(file_ext.c_str());
